--- lua/telescope-packer/init.lua
@@ -1,1 +1,0 @@
-return require("telescope._extensions.packer.main")
